add scripts for remote-build
These scripts are used by remote-build.
These versions are from my yocto-genericarm64 repo.
The pre-build.sh script saves tmp each build so the output in unique for each4746554
build.

pre-build.sh also deletes any downloads from my github which I found I needed
when force pushing to wip branches.  It should not be needed for this project
but is harmless to keep.

The post build script is only active if the DEPLOY_TFTP_DIR_BASE is defined
which is not true for this project ATM.  If defined it copys files to the
TFTP dir and runs make-image-targets in that dir.

Signed-off-by: Bill Mills <bill.mills@linaro.org>