kas: add python3 build & setup packages to sdk for xen full build
xen's 'dist' make target needs setuptools and other build tools.
Add python3-setuptools and other build packages to host side of sdk.
(Adding python3-setuptools to the image does not add it to the host side
so we don't do that.)

We also add python3-dev to the host side.  This is because
"python3 -m build" adds a -I to the python header files but uses
the host side sysroot.  This is "wrong" but the headers on both sides
will be the same if we add them.  This is how Debian cross build works.
On Debian, there is only one set of include files for all arches so
there is no "wrong" path.

Signed-off-by: Bill Mills <bill.mills@linaro.org>