kas: remove rm_work
rm_work does not really do much good.  It does its work too late to
decrease the minimum disk size needed and only preserves things that
sstate would restore anyway.  rm-work destroys valuable info on what
exactly the build did and how it did it.  This info may be needed even if
that particular recipe appears to work.  It is better to keep all the work
until it is known that everything worked.

If space is precious, it is better to do an rm -rf tmp afterword and let
sstate restore what is needed next time.  If space is not an issue it is
better to rename tmp to a per-build name or move it into a directory to
be deleted later.  Either of these is faster than rm-work.

Signed-off-by: Bill Mills <bill.mills@linaro.org>